Cleanup repo root (#16)
* Move contribution guideline documentation

Move the contribution guidelines out of the main repo and into .github
folder to clean up the repo root

* Repo Production Files

Use a `.gitattributes` file to remove things like the tests dir, the
.gitignore dir, travis config files etc. from production installs. This
means that if someone installs with `--prefer-dist`, all of these files
are excluded. This leads to faster composer installs and a cleaner vendor.